Test: add test for MeasurementSuggestions widget
