+++ test/widgets/create/measurement_suggestions_test.dart
@@ -1,0 +1,59 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:lotti/classes/journal_entities.dart';
+import 'package:lotti/database/database.dart';
+import 'package:lotti/get_it.dart';
+import 'package:lotti/themes/themes_service.dart';
+import 'package:lotti/widgets/create/suggest_measurement.dart';
+import 'package:mocktail/mocktail.dart';
+
+import '../../mocks/mocks.dart';
+import '../../test_data/test_data.dart';
+import '../../utils/measurable_utils_test.dart';
+import '../../widget_test_utils.dart';
+
+void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
+  group(' - ', () {
+    final mockJournalDb = MockJournalDb();
+    when(
+      () => mockJournalDb.watchMeasurementsByType(
+        rangeStart: any(named: 'rangeStart'),
+        rangeEnd: any(named: 'rangeEnd'),
+        type: measurableWater.id,
+      ),
+    ).thenAnswer(
+      (_) => Stream<List<JournalEntity>>.fromIterable([
+        testMeasurements(
+          <num>[111, 500, 250, 500, 250, 500, 250, 500, 100, 100, 50],
+        ),
+      ]),
+    );
+
+    setUp(() async {
+      getIt
+        ..registerSingleton<ThemesService>(ThemesService(watch: false))
+        ..registerSingleton<JournalDb>(mockJournalDb);
+    });
+    tearDown(getIt.reset);
+
+    testWidgets(
+      '',
+      (tester) async {
+        await tester.pumpWidget(
+          makeTestableWidgetWithScaffold(
+            MeasurementSuggestions(
+              measurableDataType: measurableWater,
+            ),
+          ),
+        );
+
+        await tester.pumpAndSettle();
+
+        expect(find.text('500 ml'), findsOneWidget);
+        expect(find.text('250 ml'), findsOneWidget);
+        expect(find.text('100 ml'), findsOneWidget);
+      },
+    );
+  });
+}
